requirements.txt
django==2.2.2
django-colorfield